requirements.txt
python-telegram-bot
python-dotenv
selenium
requests
# pydantic
# sqlalchemy
# fastapi
# uvicorn
# requests